Implements bookie counts for topics that should not be consumed.

+++ config.go
@@ -30,7 +30,7 @@ type event struct {
 	JSON       []map[string]interface{} `json:"json"`
 	Aggregate  bool                     `json:"aggregate"`
 	Color      string                   `json:"color"`
-	Count      int32                    `json:"count"`
+	Count      int64                    `json:"count"`
 	NoJSON     bool                     `json:"noJSON,omitempty"`
 }
